+++ tests/core/variables/test_variables.py
